feat : 추상 클래스, 매서드와 상속을 통한 확장
1. 장점
- 재사용성

2. 단점
- 구현체가 엄청 많아진다
- 다른 abstract 매서드가 추가되면 조합의 갯수가 너무 많아진다
ex : 이베이 + 보너스, 이베이 + vip, 아마존 + 보너스 ...
