+++ src/main/java/com/example/tobyspring6/payment/Client.java
@@ -1,0 +1,12 @@
+package com.example.tobyspring6.payment;
+
+import java.io.IOException;
+import java.math.BigDecimal;
+
+public class Client {
+    public static void main(String[] args) throws IOException {
+        PaymentService paymentService = new SimpleExRatePaymentService();
+        Payment payment = paymentService.prepare(100L, "USD", BigDecimal.valueOf(50.7));
+        System.out.println(payment);
+    }
+}
