+++ src/main/java/com/example/tobyspring6/payment/PaymentService.java
@@ -1,43 +1,19 @@
 package com.example.tobyspring6.payment;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.stereotype.Service;
-
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.math.BigDecimal;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.time.LocalDateTime;
-import java.util.stream.Collectors;
 
-@Service
-public class PaymentService {
+public abstract class PaymentService {
 
     public Payment prepare(Long orderId, String currency, BigDecimal foreignCurrencyAmount) throws IOException {
-        // 환율 가져오기
-        URL url = new URL("https://open.er-api.com/v6/latest/" + currency);
-        HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-        BufferedReader br = new BufferedReader(new InputStreamReader(connection.getInputStream()));
-        String response = br.lines().collect(Collectors.joining());
-        br.close();
+        BigDecimal exRate = getExRate(currency);
 
-        ObjectMapper objectMapper = new ObjectMapper();
-        ExRateData data = objectMapper.readValue(response, ExRateData.class);
-        BigDecimal exRate = data.rates().get("KRW");
-
-        // 금액 계산
         BigDecimal convertedAmount = foreignCurrencyAmount.multiply(exRate);
-
-        // 유효 시간 계산
         LocalDateTime validUntil = LocalDateTime.now().plusMinutes(30);
+
         return new Payment(orderId, currency, foreignCurrencyAmount, exRate, convertedAmount, validUntil);
     }
 
-    public static void main(String[] args) throws IOException {
-        PaymentService paymentService = new PaymentService();
-        Payment payment = paymentService.prepare(100L, "USD", BigDecimal.valueOf(50.7));
-        System.out.println(payment);
-    }
+    abstract BigDecimal getExRate(String currency) throws IOException;
 }
