+++ src/main/java/com/example/tobyspring6/payment/SimpleExRatePaymentService.java
@@ -1,0 +1,14 @@
+package com.example.tobyspring6.payment;
+
+import java.math.BigDecimal;
+
+public class SimpleExRatePaymentService extends PaymentService{
+
+    @Override
+    BigDecimal getExRate(String currency) {
+        if (currency.equals("USD")){
+            return BigDecimal.valueOf(1000);
+        }
+        throw new IllegalArgumentException("지원되지 않는 통화");
+    }
+}
