+++ src/main/java/com/example/tobyspring6/payment/WebApiExRatePaymentService.java
@@ -1,0 +1,28 @@
+package com.example.tobyspring6.payment;
+
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.math.BigDecimal;
+import java.net.HttpURLConnection;
+import java.net.URL;
+import java.util.stream.Collectors;
+
+public class WebApiExRatePaymentService extends PaymentService {
+
+    @Override
+    BigDecimal getExRate(String currency) throws IOException {
+        URL url = new URL("https://open.er-api.com/v6/latest/" + currency);
+        HttpURLConnection connection = (HttpURLConnection) url.openConnection();
+        BufferedReader br = new BufferedReader(new InputStreamReader(connection.getInputStream()));
+        String response = br.lines().collect(Collectors.joining());
+        br.close();
+
+        ObjectMapper objectMapper = new ObjectMapper();
+        ExRateData data = objectMapper.readValue(response, ExRateData.class);
+        return data.rates().get("KRW");
+    }
+}
